Protect the app
We are reading from the local storage using the
helper function readAuthToken.
If you dont have jwt token we are redirecting the client to login page.
For now we use window.location object but this intoduces performans hit.
Because we made another request to the server for the client side code.
For now we dont care.
It works for now.