Fixed A Bug
Since we set the local storage in the then block it becomes
async job and we call setSuccess outside of the then block
so it redirects the index route immediately. Since the mains stack isnt
cleared yet the local storage items set yet and we are trying to
read those items. To solve this we move the setSuccess function to the
then block.